+++ Classes/UserFunction/FormEngine/AutotranslateLanguagesItems.php
@@ -66,8 +66,8 @@ class AutotranslateLanguagesItems {
         try {
             $siteConfiguration = $siteFinder->getSiteByPageId($sitePid);
             $languages = TranslationHelper::possibleTranslationLanguages($siteConfiguration->getLanguages());
-            foreach ($languages as $language) {
-                array_push($config['items'], array($language->getTitle(), $language->getLanguageId()));
+            foreach ($languages as $languageId => $language) {
+                array_push($config['items'], array($language->getTitle(), (int)$languageId));
             }
         } catch (SiteNotFoundException $e) {