+++ Configuration/Backend/Modules.php
@@ -1,5 +1,4 @@
 <?php
-// used in TYPO3 v12
 
 use ThieleUndKlose\Autotranslate\Controller\BatchTranslationController;